+++ src/main/java/domain/model/Money.java
@@ -1,15 +1,17 @@
 package domain.model;
 
+import java.math.BigDecimal;
+
 public class Money {
     private final Currency currency;
-    private final double amount;
+    private final BigDecimal amount;
 
-    public Money(Currency currency, double amount) {
+    public Money(Currency currency, BigDecimal amount) {
         this.currency = currency;
         this.amount = amount;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 }
